Move order event SQLite repository behind repository boundary
